pyproject.toml
[tool.poetry]
authors = ["Hideki INOUE"]
description = "Discord bot with OpenAI API on AWS AppRunner"
name = "discord-bot-zundamon"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
aws-lambda-powertools = "^2.22.0"
aws-xray-sdk = "^2.12.0"
discord-py = "^2.3.2"
duckduckgo-search = "^3.8.5"
flask = "^2.3.2"
langchain = "^0.0.267"
openai = "^0.27.8"
pynacl = "^1.5.0"
python = "^3.10"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
